pyproject.toml
[project]
name = "calculator-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
